improve main layout, refactor unit tests
