+++ src/assets/styles/flex.pcss
@@ -11,5 +11,9 @@
 }
 
 .center-align {
-    justify-self: center;
+    align-self: center;
+}
+
+.center-items {
+    align-items: center;
 }
