+++ src/assets/styles/layout.pcss
@@ -3,7 +3,7 @@
 @import 'height.pcss';
 
 .header {
-    @apply .m-t-1, .m-b-1;
+    @apply .m-1;
 }
 
 .content {
@@ -12,5 +12,5 @@
 }
 
 .footer {
-    @apply .m-1, .m-b-1;
+    @apply .m-1;
 }
