+++ src/assets/styles/text.pcss
@@ -1,0 +1,34 @@
+.text-default-family {
+    font-family: "mulish";
+}
+
+.text-bold-family {
+    font-family: "mulish-bold";
+}
+
+.text-black-family {
+    font-family: "mulish-black";
+}
+
+.text-weight-normal {
+    font-weight: normal;
+}
+
+.text-weight-bold {
+    font-weight: bold;
+}
+
+.text-default {
+    @apply .text-default-family;
+    @apply .text-weight-normal;
+}
+
+.text-bold {
+    @apply .text-bold-family;
+    @apply .text-weight-normal;
+}
+
+.text-black {
+    @apply .text-black-family;
+    @apply .text-weight-normal;
+}
