+++ src/assets/styles/width.pcss
@@ -1,0 +1,120 @@
+.w-0 {
+    width: 0px;   
+}
+.w-0.5	{
+    width: 0.125rem;
+}
+.w-1	{
+    width: 0.25rem;
+}
+.w-1.5	{
+    width: 0.375rem;
+}
+.w-2 {
+    width: 0.5rem;
+}
+.w-2.5 {
+    width: 0.625rem;
+}
+.w-3 { 
+    width: 0.75rem;
+}
+.w-3.5 { 
+    width: 0.875rem;
+}
+.w-4 {
+    width: 1rem;
+}
+.w-5 {
+    width: 1.25rem;
+}
+.w-6 {
+    width: 1.5rem;
+}
+.w-7 {
+    width: 1.75rem;
+}
+.w-8 {
+    width: 2rem;
+}
+.w-9 {
+    width: 2.25rem;
+}
+.w-10 {
+    width: 2.5rem;
+}
+.w-11 {
+    width: 2.75rem;
+}
+.w-12 {
+    width: 3rem;
+}
+.w-14 {
+    width: 3.5rem;
+}
+.w-16 {
+    width: 4rem;
+}
+.w-20 {
+    width: 5rem;
+}
+.w-24 {
+    width: 6rem;
+}
+.w-28 {
+    width: 7rem;
+}
+.w-32 {
+    width: 8rem;
+}
+.w-36 {
+    width: 9rem;
+}
+.w-40 {
+    width: 10rem;
+}
+.w-44 {
+    width: 11rem;
+}
+.w-48 {
+    width: 12rem;
+}
+.w-52 { 
+    width: 13rem;
+}
+.w-56 {
+    width: 14rem;
+}
+.w-60 {
+    width: 15rem;
+}
+.w-64 {
+    width: 16rem;
+}
+.w-72 {
+    width: 18rem;
+}
+.w-80 {
+    width: 20rem;
+}
+.w-96	{
+    width: 24rem;
+}
+.w-auto	{
+    width: auto;
+}
+.w-full	{
+    width: 100%;
+}
+.w-screen	{
+    width: 100vw;
+}
+.w-min	{
+    width: min-content;
+}
+.w-max	{
+    width: max-content;
+}
+.w-fit	{
+    width: fit-content;
+}
